Filter unwanted fields of JSON & refactor package
