+++ src/main/java/zefl/TodoMain.java
@@ -1,14 +1,18 @@
 package zefl;
 
-import com.google.inject.AbstractModule;
-import com.google.inject.servlet.ServletModule;
-import com.squarespace.jersey2.guice.JerseyGuiceModule;
+import java.util.Arrays;
+import java.util.List;
+
 import org.eclipse.jetty.server.Server;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
-import java.util.List;
+import com.google.inject.AbstractModule;
+import com.google.inject.servlet.ServletModule;
+import com.squarespace.jersey2.guice.JerseyGuiceModule;
+
+import zefl.guice.TodoModule;
+import zefl.guice.JettyServerWithGuiceBuilder;
 
 public class TodoMain {
     private static final Logger LOGGER = LoggerFactory.getLogger(TodoMain.class);
