+++ src/main/java/zefl/domain/Task.java
@@ -1,4 +1,4 @@
-package zefl;
+package zefl.domain;
 
 import java.io.Serializable;
