+++ src/main/java/zefl/guice/TodoApplication.java
@@ -1,8 +1,10 @@
-package zefl;
+package zefl.guice;
 
 import org.glassfish.jersey.server.ResourceConfig;
 
 import javax.ws.rs.ApplicationPath;
+
+import zefl.TodoResource;
 
 @ApplicationPath("/")
 public class TodoApplication extends ResourceConfig {
