+++ src/main/java/zefl/guice/TodoModule.java
@@ -1,9 +1,15 @@
-package zefl;
+package zefl.guice;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.name.Named;
 import com.google.inject.name.Names;
+
+import zefl.TodoResource;
+import zefl.dao.TodoDAO;
+import zefl.dao.TodoSimplePersistentDaoImpl;
+import zefl.service.TodoService;
+import zefl.service.TodoServiceImpl;
 
 public class TodoModule extends AbstractModule {
