+++ src/test/java/zefl/TestTodoModule.java
@@ -4,6 +4,11 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.name.Named;
 
+import zefl.dao.TodoDAO;
+import zefl.dao.TodoMemDAOImpl;
+import zefl.service.TodoService;
+import zefl.service.TodoServiceImpl;
+
 public class TestTodoModule extends AbstractModule {
 
     @Override
