+++ src/test/java/zefl/TodoResourceTest.java
@@ -22,6 +22,10 @@ import java.util.Map;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
+import zefl.domain.Task;
+import zefl.domain.Todo;
+import zefl.service.TodoService;
+
 public class TodoResourceTest extends JerseyTest {
     @Override
     protected Application configure() {
