+++ src/test/java/zefl/UtilTest.java
@@ -9,6 +9,10 @@ import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 
+import zefl.domain.Task;
+import zefl.domain.Todo;
+import zefl.util.Util;
+
 public class UtilTest {
     @Test
     public void getFormattedStringOf() throws Exception {
